+++ lib/broadcast.rb
@@ -22,6 +22,7 @@ require_relative 'broadcast/resources/autopilots'
 require_relative 'broadcast/resources/discovery'
 require_relative 'broadcast/resources/suppressions'
 require_relative 'broadcast/resources/global_suppressions'
+require_relative 'broadcast/resources/channel'
 require_relative 'broadcast/resources/migration'
 
 # ActionMailer integration — only loaded when Rails is present